+++ pages/test/index.tsx
@@ -63,7 +63,9 @@ export default function Test() {
         <tbody>
           <tr>
             <td className={commonCellClass}>inputfield</td>
-            <td className={commonRowClass}>{/* <Signup /> */}</td>
+            <td className={commonRowClass}>
+              <Signup />
+            </td>
           </tr>
           <tr className="border-b border-gray-300">
             <td className={commonCellClass}>Button</td>
@@ -98,7 +100,9 @@ export default function Test() {
               <LinkBar link="https://www.google.com" />
 
               <td className={commonCellClass}>inputfield</td>
-              <td className={commonRowClass}>{/* <Signup /> */}</td>
+              <td className={commonRowClass}>
+                <Signup />
+              </td>
             </td>
           </tr>
           <tr className="border-b border-gray-300">
